backend/chatbotbackend/requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
pymongo
langchain
langchain-community
mysql
pip install mysql-connector-python
